requirements.txt
scipy==1.14.0
selenium==4.23.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
threadpoolctl==3.5.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
python-telegram-bot==13.7
pandas==1.3.3
scikit-learn==0.24.2
requests==2.26.0
